O-Lens/backend/requirements.txt
fastapi
tensorflow
Pillow
uvicorn
python-multipart